+++ meal-ui/src/styles/local/partials/5-pages/home-page/home-page.css
@@ -9,3 +9,7 @@
 	padding-left: var(--ns-full);
 	padding-right: var(--ns-full);
 }
+
+.quick-recipes-view > div:last-child {
+	margin-top: var(--ns-full);
+}
